cleaned some stuff up
