+++ src/main/java/com/maastricht/university/logic/game/TileGroup.java
@@ -14,8 +14,8 @@ public class TileGroup {
     public TileGroup(LogicTile member) {
         this.playerColor = member.getPlayerColor();
         this.members = new LinkedList<>();
-        members.add(member);
 
+        members.add(member);
         groupID = count++;
     }
 
@@ -30,37 +30,11 @@ public class TileGroup {
     }
 
     /**
-     * Returns the ID of the group
-     * @return the ID of the group
-     */
-    public int getGroupID() {
-        return groupID;
-    }
-
-    /**
      * Returns all the LogicTiles in this group
      * @return all the LogicTiles in this group
      */
     public List<LogicTile> getMembers() {
-        return new LinkedList<LogicTile>(members);
-    }
-
-    /**
-     * Resets the list of LogicTiles in this group
-     * @param members the new list of LogicTiles
-     */
-    public void setMembers(List<LogicTile> members) {
-        try {
-            if (!membersAreSameColor(members))
-                throw new IllegalArgumentException("Members are not of the same PlayerColor");
-            if (members.get(0).getPlayerColor() != playerColor)
-                throw new IllegalArgumentException("Cant set tilelist as members of this group because the tile's playerColor do not match the groups playerColor");
-
-            this.members = members;
-
-        } catch (Exception e) {
-            System.out.println(e);
-        }
+        return new LinkedList<>(members);
     }
 
     /**
@@ -77,27 +51,6 @@ public class TileGroup {
         } catch (Exception e) {
             System.out.println(e);
         }
-    }
-
-
-
-
-
-
-
-
-
-
-
-
-    /**
-     * checks if all members in a list are of the same playerColor
-     */
-    private boolean membersAreSameColor(List<LogicTile> members) {
-        for (LogicTile member : members)
-            if (member.getPlayerColor() != playerColor)
-                return false;
-        return false;
     }
 
     @Override
